backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
mysql-connector-python
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
python-dotenv
